Migrate _app.jsx and _document.jsx to layout.tsx
